pyproject.toml
[tool.poetry]
name = "clickhouse"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
clickhouse-connect = "^0.5.21"
python-dotenv = "^1.0.0"
numpy = "^1.24.3"
seaborn = "^0.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
